Adressing code review comments and added a undefined condition for blob storage no file case
